Updated module structure and Readme
