+++ src/bookscraper/__init__.py
@@ -1,2 +1,2 @@
 from .scrape_details import scrape_amazon
-from .book_utils import hash_book, extract_year_from_date
+from .book_utils import hash_book, extract_year_from_date, print_log
